requirements.txt
awswrangler==3.11.0
boto3==1.37.26
pandas==2.2.3
python-dotenv==1.1.0
